seprate store\index.js
