+++ src/pages/PageCategory.vue
@@ -1,11 +1,12 @@
 <template>
-    <div class="col-full">
+    <div v-if="category" class="col-full">
         <h1>{{category.name}}</h1>
         <CategoryListItem :category="category"/>
     </div>
 </template>
 
 <script>
+import {mapActions} from 'vuex'
 import CategoryListItem from '@/components/CategoryListItem'
 // import sourceData from '@/data'
 export default {
@@ -22,6 +23,15 @@ export default {
     category () {
       return this.$store.state.categories[this.id]
     }
+  },
+  methods: {
+    ...mapActions(['fetchCategory', 'fetchForums'])
+  },
+  created () {
+    this.fetchCategory({id: this.id})
+      .then(category => {
+        this.fetchForums({ids: category.forums})
+      })
   }
 }
 </script>
