+++ src/pages/PageThreadCreate.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="col-full push-top">
+    <div v-if="forum" class="col-full push-top">
 
           <h1>Create new thread in <i>{{forum.name}}</i></h1>
 
@@ -12,6 +12,7 @@
 </template>
 
 <script>
+import {mapActions} from 'vuex'
 import ThreadEditor from '@/components/ThreadEditor'
 export default {
   components: {
@@ -36,8 +37,9 @@ export default {
     }
   },
   methods: {
+    ...mapActions(['createThread', 'fetchForum']),
     save ({title, text}) {
-      this.$store.dispatch('createThread', {
+      this.createThread({
         forumId: this.forum['.key'],
         title,
         text
@@ -48,6 +50,9 @@ export default {
     cancel () {
       this.$router.push({name: 'Forum', params: {id: this.forum['.key']}})
     }
+  },
+  created () {
+    this.fetchForum({id: this.forumId})
   }
 }
 </script>
