+++ src/pages/PageThreadEdit.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="col-full push-top">
+    <div v-if="thread && text" class="col-full push-top">
 
           <h1>Editing <i>{{thread.title}}</i></h1>
 
@@ -14,6 +14,7 @@
 </template>
 
 <script>
+import {mapActions} from 'vuex'
 import ThreadEditor from '@/components/ThreadEditor'
 export default {
   components: {
@@ -30,12 +31,15 @@ export default {
       return this.$store.state.threads[this.id]
     },
     text () {
-      return this.$store.state.posts[this.thread.firstPostId].text
+      const post = this.$store.state.posts[this.thread.firstPostId]
+      return post ? post.text : null
     }
   },
   methods: {
+    ...mapActions(['updateThread', 'fetchPost', 'fetchThread']),
     save ({title, text}) {
-      this.$store.dispatch('updateThread', {
+      // ORIGINAL >> this.$store.dispatch('updateThread', {
+      this.updateThread({
         id: this.id,
         title,
         text
@@ -46,6 +50,11 @@ export default {
     cancel () {
       this.$router.push({name: 'ThreadShow', params: {id: this.id}})
     }
+  },
+  created () {
+    // OROGINAL this.$store.dispatch('fetchThread', {id: this.id})
+    this.$store.dispatch('fetchThread', {id: this.id})
+      .then(thread => this.fetchPost({id: thread.firstPostId}))
   }
 }
 </script>
