+++ src/pages/PagethreadShow.vue
@@ -22,6 +22,7 @@
 </template>
 <script>
 // import firebase from 'firebase'
+import {mapActions} from 'vuex'
 import PostList from '@/components/PostList'
 import PostEditor from '@/components/PostEditor'
 import {countObjectProperties} from '@/utils/index'
@@ -68,6 +69,9 @@ export default {
       return countObjectProperties(this.thread.contributors)
     }
   },
+  methods: {
+    ...mapActions(['fetchThread', 'fetchUser', 'fetchPosts'])
+  },
   beforeCreate () {
     // this.id = undefined
   },
@@ -99,14 +103,14 @@ export default {
     // })
 
     // fetch thread
-    this.$store.dispatch('fetchThread', {id: this.id})
+    this.fetchThread({id: this.id})
       .then(thread => {
         // fetch user
-        this.$store.dispatch('fetchUser', {id: thread.userId})
-        this.$store.dispatch('fetchPosts', {ids: Object.keys(thread.posts)})
+        this.fetchUser({id: thread.userId})
+        this.fetchPosts({ids: Object.keys(thread.posts)})
         .then(posts => {
           posts.forEach(post => {
-            this.$store.dispatch('fetchUser', {id: post.userId})
+            this.fetchUser({id: post.userId})
           })
         })
